zip_money_backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
django-graphql-auth==0.3.7
django-graphql-jwt==0.4.0
django-graphql-social-auth==0.1.4
flake8==6.1.0
graphene==3.3
graphene-django==3.1.5
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.6
isort==5.13.1
mccabe==0.7.0
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
promise==2.3
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
xhtml2pdf
https://github.com/instruct-br/graphene-django-pagination/archive/master.zip
